+++ elmyra.ip.access.epo/elmyra/ip/access/epo/static/js/app/models/sip.js
@@ -1,0 +1,6 @@
+// -*- coding: utf-8 -*-
+// (c) 2014 Andreas Motl, Elmyra UG
+
+SipSearch = DatasourceSearch.extend({
+    url: '/api/ftpro/published-data/search',
+});